The visual system is working
The guard has a cone field of view with indirect side view where it can notice stuff happening.
If it sees the player it will try to follow and if target is lost it will finish walking to last known location and then return to patrol.
Walls can block vision if on a layer included on the block list.
Potential targets need to be included on the vision list.